+++ Chemistry/tests/DataModelTests/OrbitalIntegralTests/OrbitalIntegralTests.cs
@@ -26,9 +26,9 @@ namespace Microsoft.Quantum.Chemistry.Tests
         [InlineData(0, 0, 1, 2, 8)]
         [InlineData(0, 1, 2, 0, 4)]
         [InlineData(0, 1, 2, 3, 8)]
-        public void OrbitalIntegralEnumerateOrbitalSymmetriesTest(Int64 i, Int64 j, Int64 k, Int64 l, Int64 elements)
+        public void OrbitalIntegralEnumerateOrbitalSymmetriesTest(int i, int j, int k, int l, int elements)
         {
-            OrbitalIntegral orbitalIntegral = new OrbitalIntegral(new Int64[] { i, j, k, l });
+            OrbitalIntegral orbitalIntegral = new OrbitalIntegral(new int[] { i, j, k, l });
             var orbitalIntegrals = orbitalIntegral.EnumerateOrbitalSymmetries();
             Assert.Equal(elements, orbitalIntegrals.Length);
         }
@@ -44,9 +44,9 @@ namespace Microsoft.Quantum.Chemistry.Tests
         [InlineData(0, 0, 1, 2, 8)]
         [InlineData(0, 1, 2, 0, 4)]
         [InlineData(0, 1, 2, 3, 8)]
-        public void OrbitalIntegralEnumerateSpinOrbitalsTest(Int64 i, Int64 j, Int64 k, Int64 l, Int64 elements)
+        public void OrbitalIntegralEnumerateSpinOrbitalsTest(int i, int j, int k, int l, int elements)
         {
-            OrbitalIntegral orbitalIntegral = new OrbitalIntegral(new Int64[] { i, j, k, l });
+            OrbitalIntegral orbitalIntegral = new OrbitalIntegral(new int[] { i, j, k, l });
             var orbitalIntegrals = orbitalIntegral.EnumerateOrbitalSymmetries();
             var spinOrbitals = orbitalIntegrals.EnumerateSpinOrbitals();
             Assert.Equal(elements * 4, spinOrbitals.Length);
